Share PortableThreadPool.WorkerThread common code
Share the code between non-browser implementations and
browser+threads.

The differences are just in how the work loop is started and implemented
